requirements.txt
flask==2.0.1
PyMuPDF==1.18.17
opencv-python-headless==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
werkzeug==2.0.1